installation of bson_ext for non-java rubies
Applying this change allows fl00r/em-mongo to work on JRuby.

The bson_ext provides C extensions which are not supported for java
platforms (i.e. JRuby). This change will install bson_ext only if
the gem is installed on a non-java ruby.

This implemented using an extconf.rb file.